+++ runner/src/server/plugins/engine/services/addressService.ts
@@ -1,4 +1,4 @@
-import { getJson } from "./http-service";
+import { getJson } from "./httpService";
 
 async function findByPostcode(key, postcode) {
   const findByPostcodeUrl = `https://api.ordnancesurvey.co.uk/places/v1/addresses/postcode?lr=EN&fq=logical_status_code:1&dataset=DPA&postcode=${postcode}&key=${key}`;
